Change db file so accesses dev DB
I plan to convert this file so we rely on environment variables to
store the password, rather than plain text in a GitHub repo! Also
ideally the solution will dynamically context switch between dev,
test and propd without any manual file editing of the deployed code.

That's all what I plan to do but for now the more modest aim of
connecting to the DB that's actually there has been put in place.