Removed temp file
This file was a hang-over from when auto-quiz was implemented in the last
tournament. Because it was implemented halfway through the tournament this
script was written to bulk process the previous results. This is no longer
needed as AutoQuiz will be calculated normally as each new match comes in now.