Fix Mr Mode
Previous code was just taking the first instance of the score predicted with
maximal frequency. This is fine if there's only one. If there is more than one
though the other predictions, which are equally correct, are ignored.

This fix filters down onto all the maximal frequency predictions and then
averages them (median!) to get a better mode.