Fix Mr Median
The median algorithm was previously wrong as it referenced the wrong elements.
As arrays are zero based in PHP we actually need to reference the element below
the one previously takereferenced the wrong elements. As arrays are zero based
in PHP we actually need to reference the element below the one previously taken