Updated scoring for missed predictions
Missed predictions are scored as though the player predicted 0 - 0 but with the
further limitation that they cannot get any positive points as a consequence of
this assumed prediction. The limitation is applied per prediction strand. Had to
make some tweaks to the results e-mail to to better show these new results.